% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vPro.R
\name{annotate_peptides}
\alias{annotate_peptides}
\title{Annotate identified peptides with database information and
get information about the occurrences of the peptides in the
measurement and the peptide library. Calculate the virID score.}
\usage{
annotate_peptides(pep, pepLib, nr_human_peps = 591159, peptide_fdr = 0.01)
}
\arguments{
\item{pep}{measured peptides (data frame)}

\item{nr_human_peps}{number of human peptides in Library}

\item{peptide_fdr}{false discovery rate that should be applied}

\item{pep_lib}{vPro Viral Peptide Library (data frame)}
}
\value{
annotated identified peptides (data frame)
}
\description{
Annotate identified peptides with database information and
get information about the occurrences of the peptides in the
measurement and the peptide library. Calculate the virID score.
}
\examples{
annotatedPeptides <- annotate_peptides(diann, virLib)
}
