% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vPro.R
\name{assign_species}
\alias{assign_species}
\title{Assign species taxonomy to all measurements. Apply filters and aggregate
functions.}
\usage{
assign_species(
  annPep,
  min_peptides_per_species = 2,
  min_vProIDScore = 2,
  topn_precursor = 3,
  filter_vProIDScore = TRUE
)
}
\arguments{
\item{annPep}{annotated measured peptides and their vProID Score (data frame)}
}
\value{
species information for each measurement (data frame)
}
\description{
Assign species taxonomy to all measurements. Apply filters and aggregate
functions.
}
\examples{
run2species <- assign_species(annotatedPeptides)
}
