% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vPro.R
\name{assign_subspecies}
\alias{assign_subspecies}
\title{Assign subspecies taxonomy.}
\usage{
assign_subspecies(ann_pep, ann_pep_tax, min_peptides_per_subspecies = 2)
}
\arguments{
\item{annPep}{annotated measured peptides and their virID Score (data frame)}

\item{annPep_tax}{species information for each measurement (data frame)}
}
\value{
species and subspecies information for each measurement (data frame)
}
\description{
Assign subspecies taxonomy.
}
\examples{
run2species <- assign_subspecies(annotatedPeptides, run2species)
}
