% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vPro.R
\name{assign_viral_species}
\alias{assign_viral_species}
\title{Assign taxonomy to each run using a virusID Score}
\usage{
assign_viral_species(
  file_peptides,
  file_virusDB,
  file_export,
  min_cScore = 0.95,
  nr_human_peptides = 591159,
  fdr = 0.01,
  min_pep_species = 2,
  min_pep_subspecies = 2,
  min_vProIDScore = 2,
  topn_precursor = 3,
  filter_vProIDScore = TRUE
)
}
\arguments{
\item{file_peptides}{DiaNN main output file containing the identified precursors
\itemize{
\item tab-delimited text file
\item header included
\item needs to contain columns: Run, Stripped.Sequence, Precursor.Normalised,
CScore
}}

\item{file_virusDB}{vPro Peptide Library
\itemize{
\item tab-delimited text file
\item header included
\item needs to contain columns: Species, Subspecies, Proteomes, PeptideSequence
}}

\item{file_export}{Output file of vPro-MS Virus Identification
\itemize{
\item tab-delimited text file
\item header included
\item contains columns: Run, Species, Subspecies, No.Peptide.Sequences,
No.Peptide.Sequences.Subspecies,    vProID.Score,
Virus.Quantity, Peptide.Sequences, CScores,    Proteomes
}}

\item{min_cScore}{minimal CScore threshold for precursor identification}

\item{nr_human_peptides}{Number of human peptides in the library}

\item{fdr}{FalseDiscoveryRate for peptide identification}

\item{min_pep_species}{minimal number of identified peptides necessary for a
species assignment}

\item{min_pep_subspecies}{minimal number of identified peptides necessary for
a subspecies assignment}

\item{min_vProIDScore}{minimal vProID score necessary for taxonomic assignment}

\item{topn_precursor}{number of most abundant precursor that are
used for quantification}

\item{filter_vProIDScore}{should the results be filtered according to vProIDScore
(Boolean)}
}
\description{
Assign taxonomy to each run using a virusID Score
}
\examples{
assign_viral_species("diann_output.tsv", "vPro.Peptide.Library.tsv", "results_vPro.tsv")
}
