% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vPro.R
\name{export_results}
\alias{export_results}
\title{Write results to a file.}
\usage{
export_results(results, f_name = file_export)
}
\arguments{
\item{results}{Results including species information (data frame)}

\item{f_name}{File name for output (file path)}
}
\value{
None
}
\description{
Write results to a file.
}
\examples{
export_results(run2species, f_name = "results_vPro.tsv")
}
