% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vPro.R
\name{read_peptides}
\alias{read_peptides}
\title{Load a DIA-NN report file}
\usage{
read_peptides(f_name, min_cscore = 0.95)
}
\arguments{
\item{f_name}{A path to the DIA-NN main report file.}

\item{min_cscore}{Filter entries below this CScore. (default 0.95)}
}
\value{
Detected Peptides (data frame).
}
\description{
Load a DIA-NN report file
}
\examples{
diann <- read_peptides(f_name = "diann-report.tsv", min_cscore = 0.9)
diann <- read_peptides("diann-report.tsv")
}
