% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vPro.R
\name{read_virus_db}
\alias{read_virus_db}
\title{Load the vPro Peptide Library}
\usage{
read_virus_db(f_name)
}
\arguments{
\item{f_name}{A path to the vPro Peptide Library.}
}
\value{
vPro Peptide Library (data frame)
}
\description{
Load the vPro Peptide Library
}
\examples{
virLib <- read_virus_db(f_name = "vPro.Peptide.Library.txt")
virLib <- read_virus_db("vPro.Peptide.Library.txt")
}
